requirements.txt
openai
langchain
huggingface_hub
streamlit
python-dotenv